requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
discord.py==1.2.3
idna==2.8
multidict==4.5.2
numpy==1.17.2
pandas==0.25.1
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
websockets==6.0
yarl==1.3.0
